requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.17
python-dotenv==1.0.0
unstructured==0.12.0
sentence-transformers==2.2.2
openai==1.3.0
faiss-cpu==1.7.4
